requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0
beautifulsoup4>=4.12.0
supabase>=2.3.0
python-dotenv>=1.0.0
